+++ app/dashboard/components/AnimatedCounter.tsx
@@ -1,0 +1,63 @@
+'use client'
+
+import { useState, useEffect, useRef, memo } from 'react'
+
+interface AnimatedCounterProps {
+  value: number
+  duration?: number
+}
+
+function AnimatedCounter({ value, duration = 2000 }: AnimatedCounterProps) {
+  const [count, setCount] = useState(0)
+  const [isVisible, setIsVisible] = useState(false)
+  const ref = useRef<HTMLSpanElement>(null)
+
+  useEffect(() => {
+    const observer = new IntersectionObserver(
+      ([entry]) => {
+        if (entry.isIntersecting && !isVisible) {
+          setIsVisible(true)
+        }
+      },
+      { threshold: 0.1 }
+    )
+
+    if (ref.current) {
+      observer.observe(ref.current)
+    }
+
+    return () => observer.disconnect()
+  }, [isVisible])
+
+  useEffect(() => {
+    if (!isVisible) return
+
+    let startTime: number | null = null
+    const startValue = 0
+    const endValue = value
+
+    const animate = (currentTime: number) => {
+      if (!startTime) startTime = currentTime
+      const progress = Math.min((currentTime - startTime) / duration, 1)
+      
+      // Easing function for smooth animation
+      const easeOutQuart = 1 - Math.pow(1 - progress, 4)
+      const currentValue = Math.floor(startValue + (endValue - startValue) * easeOutQuart)
+      
+      setCount(currentValue)
+      
+      if (progress < 1) {
+        requestAnimationFrame(animate)
+      } else {
+        setCount(endValue)
+      }
+    }
+
+    requestAnimationFrame(animate)
+  }, [isVisible, value, duration])
+
+  return <span ref={ref}>{count.toLocaleString()}</span>
+}
+
+export default memo(AnimatedCounter)
+
